重构摘要算法代码
