+++ src/main/java/io/xjar/XJdkEncryptor.java
@@ -47,11 +47,11 @@ public class XJdkEncryptor implements XEncryptor {
             Cipher cipher = Cipher.getInstance(algorithm);
             cipher.init(Cipher.ENCRYPT_MODE, new SecretKeySpec(key.getEncryptKey(), algorithm));
             cis = new CipherInputStream(in, cipher);
-            XKit.transfer(cis, out);
+            XTool.transfer(cis, out);
         } catch (Exception e) {
             throw new IOException(e);
         } finally {
-            XKit.close(cis);
+            XTool.close(cis);
         }
     }
