+++ src/main/java/io/xjar/boot/XBootEncryptor.java
@@ -97,7 +97,7 @@ public class XBootEncryptor extends XArchiveEncryptor<JarArchiveEntry> implement
                     }
                     XEncryptor encryptor = filtered ? xEncryptor : xNopEncryptor;
                     try (OutputStream eos = encryptor.encrypt(key, nos)) {
-                        XKit.transfer(nis, eos);
+                        XTool.transfer(nis, eos);
                     }
                 }
                 // BOOT-INF/lib/**
@@ -112,14 +112,14 @@ public class XBootEncryptor extends XArchiveEncryptor<JarArchiveEntry> implement
                     jar.setCrc(cos.getChecksum().getValue());
                     zos.putArchiveEntry(jar);
                     ByteArrayInputStream bis = new ByteArrayInputStream(bos.toByteArray());
-                    XKit.transfer(bis, nos);
+                    XTool.transfer(bis, nos);
                 }
                 // OTHER
                 else {
                     JarArchiveEntry jarArchiveEntry = new JarArchiveEntry(entry.getName());
                     jarArchiveEntry.setTime(entry.getTime());
                     zos.putArchiveEntry(jarArchiveEntry);
-                    XKit.transfer(nis, nos);
+                    XTool.transfer(nis, nos);
                 }
                 zos.closeArchiveEntry();
             }
@@ -147,8 +147,8 @@ public class XBootEncryptor extends XArchiveEncryptor<JarArchiveEntry> implement
 
             zos.finish();
         } finally {
-            XKit.close(zis);
-            XKit.close(zos);
+            XTool.close(zis);
+            XTool.close(zos);
         }
     }
